+++ esdglider/glider.py
@@ -301,7 +301,7 @@ def binary_to_nc(
     sci_timeseries_pyglider : bool, default True
         Should the function use pyglider.slocum.binary_to_timeseries to create
         create the science timeseries (True),
-        or glider.timeseries_raw_to_sci (False)
+        or glider.raw_to_sci_timeseries (False)
     gridded_depth_measured : bool, default False
         Should the function pass the science timeseries directly to 
         pyglider.ncprocess.make_gridfiles (False), 
@@ -456,7 +456,7 @@ def binary_to_nc(
 
         else:
             _log.info("Generating science timeseries, via raw_to_sci_timeseries")
-            outname_tssci = timeseries_raw_to_sci(
+            outname_tssci = raw_to_sci_timeseries(
                 outname_tsraw,
                 tsdir,
                 deploymentyaml,
@@ -1359,7 +1359,7 @@ def binary_to_raw_timeseries(
     return outname
 
 
-def timeseries_raw_to_sci(
+def raw_to_sci_timeseries(
     inname,
     outdir,
     deploymentyaml,
@@ -1368,8 +1368,8 @@ def timeseries_raw_to_sci(
     maxgap=300,
 ):
     """
-    Go from raw timeseries (from esdglider.glider.binary_to_raw)
-    to processed science timeseries.
+    Go from raw timeseries (from esdglider.glider.binary_to_raw_timeseries)
+    to a processed science timeseries.
     This function can be used in cases where different science sensors are
     on at different times, e.g. PAM deployments, and thus it is not possible
     to get the full science timeseries using dbdreader.get_sync.